Remove 1.18.2 from supported version list. (#2074)
* Remove 1.18.2 from supported version list.

1.18.2 support was removed with changes in commit 056cff4b6f630fdc5a5b58b3dcb70434511757cf

Also, mark 1.19, 1.19.1 and 1.19.2 as incompatible. It happens because of GSon library changes that prevents bentobox compiled with Spigot 1.19.3 to run on older versions.

* Fixes GSON crashes introduced by migrating to 1.19.3 api.

+++ src/main/java/world/bentobox/bentobox/database/json/adapters/LocationTypeAdapter.java
@@ -24,8 +24,9 @@ public class LocationTypeAdapter extends TypeAdapter<Location> {
         out.value(location.getX());
         out.value(location.getY());
         out.value(location.getZ());
-        out.value(location.getYaw());
-        out.value(location.getPitch());
+        // This is required for 1.19-1.19.2 compatibility.
+        out.value((double) location.getYaw());
+        out.value((double) location.getPitch());
         out.endArray();
     }
